Rename to `appActions.js`
