+++ e2e/tests/plugins/imagery/exampleImagery.e2e.spec.js
@@ -28,7 +28,7 @@ but only assume that example imagery is present.
 
 const { test } = require('../../../fixtures.js');
 const { expect } = require('@playwright/test');
-const { waitForAnimations } = require('../../../commonActions.js');
+const { waitForAnimations } = require('../../../appActions.js');
 
 const backgroundImageSelector = '.c-imagery__main-image__background-image';
